task 1 finished

+++ src/main/java/fu/infb/ue09/klassenhierarchie/Geom3D.java
@@ -1,6 +1,6 @@
 package fu.infb.ue09.klassenhierarchie;
 
 public abstract class Geom3D {
-	public abstract int volume();
-	public abstract int surfaceArea();
+	public abstract double volume();
+	public abstract double surfaceArea();
 }
